Widen surface-cell count API to 64-bit to avoid int overflow
get_number_of_surface_cells() returned int (truncating surfaceCellVec_.size()),
so a freeze-out surface exceeding 2^31 cells would overflow. Return long and
index get_surface_cell_with_index() with long. Harmless for typical events but
removes a latent overflow in large-scale production.

+++ src/evolve.h
@@ -46,8 +46,8 @@ class Evolve {
     ~Evolve() {clearSurfaceCellVector();}
 
     void clearSurfaceCellVector() {surfaceCellVec_.clear();}
-    int get_number_of_surface_cells() const {return(surfaceCellVec_.size());}
-    void get_surface_cell_with_index(const int i, SurfaceCell &cell_i) {
+    long get_number_of_surface_cells() const {return(surfaceCellVec_.size());}
+    void get_surface_cell_with_index(const long i, SurfaceCell &cell_i) {
         cell_i = surfaceCellVec_[i];
     }
